django/requirements.txt
# Django backend 
gunicorn==22.0.0 
djangorestframework-simplejwt==5.3.1

drf-spectacular
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
djangorestframework==3.15.2
idna==3.10
numpy==2.2.4
pillow==11.1.0
psycopg2-binary==2.9.10
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
 

# ML dependencies were removed becase i can not remember why i put them in this container
# opencv-python
# tensorflow
# torch
# numpy
# scipy
# matplotlib
